#161 - added a couple of things
