+++ .github/workflows/build.yml
@@ -23,6 +23,7 @@ jobs:
     - run: npm --prefix users/userservice test -- --coverage
     - run: npm --prefix llmservice test -- --coverage
     - run: npm --prefix gatewayservice test -- --coverage
+    - run: npm --prefix gameservice test -- --coverage
     - run: npm --prefix webapp test -- --coverage
     - name: Analyze with SonarQube
       uses: SonarSource/sonarqube-scan-action@master
